requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
pyactiveresource==2.2.2
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
ShopifyAPI==12.4.0
six==1.16.0
urllib3==2.2.1
